refactor: separate switchLocalePath SSR into plugin
